bug fix and error handling

+++ Build/weeklyRestore.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-#RESTORE
+#WEEKLY RESTORE
 
 if [[ -d $1 ]] || [[ -z $1 ]]
 then
@@ -7,18 +7,25 @@ then
 	exit;
 fi
 
-read -p "this will erase all current data, are you sure to continue ?" yn
+
+p=`echo "$1" | grep -c "wb"`
+if [[ $p != 1 ]]
+then
+	echo "must be weekly backup file !"
+	exit;
+fi
+
+read -p "This is a weekly restore. It will set the db to the backup's state. Are you sure to continue ?" yn
 
 tar -xzvf "$1"
 dir=$(echo $1 | sed 's/\.tar\.gz//');
 
 case $yn in 
 	[Yy]* ) 
-		rm -rf Data/;
 		mongo --quiet --eval 'db.getMongo().getDBNames().forEach(function(i){db.getSiblingDB(i).dropDatabase()})';;
 	[Nn]* ) exit;;
 esac
 
 mongorestore --host 127.0.0.1 --port 27017 --dir "$dir"
-mv "$dir/Data" Data
+mv -vf "$dir/file/"* Data/file/
 rm -rf "$dir"
